requirements.txt
discord.py>=1.3.1
jishaku==1.17.0.169
sr_api
